pyproject.toml
[project]
name = "faiss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.4",
    "langchain-text-splitters>=1.1.0",
    "langchain[openai]>=1.2.0",
    "langgraph>=1.0.5",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
]
